refactor: 점수 계산 로직 이동 - 점수 계산을 chessBoard -> result로 이동 - 출력 수정 - 테스트 코드 수정
