+++ src/main/java/chess/view/OutputView.java
@@ -15,7 +15,7 @@ public class OutputView {
     public static final String MOVE_COMMAND = "> 게임 이동 : move source위치 target위치 - 예. move b2 b3";
     public static final String GAME_END = "> 체스 게임이 종료되었습니다.";
     public static final String STATUS_COMMAND = "> 결과 확인: status";
-    public static final String RESULT_FORMAT = "%s: %.1f점 - %s%n";
+    public static final String RESULT_FORMAT = "%s: %.1f점";
 
     private OutputView() {
     }
@@ -36,12 +36,8 @@ public class OutputView {
     }
 
     public static void printResult(Result result) {
-        Map<Color, Double> score = result.getResult();
-        Map<Color, String> winOrLose = result.getWinOrLose();
-        System.out.printf(RESULT_FORMAT, Color.BLACK.name(), score.get(Color.BLACK),
-            winOrLose.get(Color.BLACK));
-        System.out.printf(RESULT_FORMAT, Color.WHITE.name(), score.get(Color.WHITE),
-            winOrLose.get(Color.WHITE));
+        System.out.printf(RESULT_FORMAT, Color.BLACK.name(), result.score(Color.BLACK));
+        System.out.printf(RESULT_FORMAT, Color.WHITE.name(), result.score(Color.WHITE));
     }
 
     public static void gameEnd() {
