+++ src/test/java/chess/ResultTest.java
@@ -2,41 +2,53 @@ package chess;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.domain.board.ChessBoard;
+import chess.domain.board.Position;
 import chess.domain.game.Result;
+import chess.domain.piece.Bishop;
 import chess.domain.piece.Color;
+import chess.domain.piece.King;
+import chess.domain.piece.Knight;
+import chess.domain.piece.Pawn;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Queen;
+import chess.domain.piece.Rook;
 import java.util.Map;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 public class ResultTest {
 
-    @Test
-    @DisplayName("승패 결과 확인, WHITE 승리")
-    void resultWhite() {
-        Result result = new Result(10, 15);
-        Map<Color, String> winOrLose = result.getWinOrLose();
+    ChessBoard chessBoard;
 
-        assertThat(winOrLose.get(Color.BLACK)).isEqualTo(Result.LOSE);
-        assertThat(winOrLose.get(Color.WHITE)).isEqualTo(Result.WIN);
+    @BeforeEach
+    void setUp() {
+        chessBoard = new ChessBoard();
+        Map<Position, Piece> board = chessBoard.getChessBoard();
+        board.put(Position.of("b8"), new King(Color.BLACK));
+        board.put(Position.of("c8"), (new Rook(Color.BLACK)));
+        board.put(Position.of("a7"), (new Pawn(Color.BLACK)));
+        board.put(Position.of("c7"), (new Pawn(Color.BLACK)));
+        board.put(Position.of("d7"), (new Bishop(Color.BLACK)));
+        board.put(Position.of("b6"), (new Pawn(Color.BLACK)));
+        board.put(Position.of("e6"), (new Queen(Color.BLACK)));
+
+        board.put(Position.of("f4"), (new Knight(Color.WHITE)));
+        board.put(Position.of("g4"), (new Queen(Color.WHITE)));
+        board.put(Position.of("f3"), (new Pawn(Color.WHITE)));
+        board.put(Position.of("h3"), (new Pawn(Color.WHITE)));
+        board.put(Position.of("f2"), (new Pawn(Color.WHITE)));
+        board.put(Position.of("g2"), (new Pawn(Color.WHITE)));
+        board.put(Position.of("e1"), (new Rook(Color.WHITE)));
+        board.put(Position.of("f1"), (new King(Color.WHITE)));
     }
 
     @Test
-    @DisplayName("승패 결과 확인, 무승부")
-    void resultDraw() {
-        Result result = new Result(15, 15);
-        Map<Color, String> winOrLose = result.getWinOrLose();
-
-        assertThat(winOrLose.get(Color.BLACK)).isEqualTo(Result.DRAW);
-        assertThat(winOrLose.get(Color.WHITE)).isEqualTo(Result.DRAW);
-    }
-
-    @Test
-    @DisplayName("승패 결과 확인, BLACK 승리")
-    void resultBlack() {
-        Result result = new Result(15, 10);
-        Map<Color, String> winOrLose = result.getWinOrLose();
-
-        assertThat(winOrLose.get(Color.BLACK)).isEqualTo(Result.WIN);
-        assertThat(winOrLose.get(Color.WHITE)).isEqualTo(Result.LOSE);
+    @DisplayName("각 진영 별 점수 계산")
+    void scoreTest() {
+        Result result = new Result(chessBoard.getChessBoard());
+        assertThat(result.score(Color.WHITE)).isEqualTo(19.5);
+        assertThat(result.score(Color.BLACK)).isEqualTo(20.0);
     }
 }
